allow up to 2 decimal places for order quantity
Updated the order quantity input fields to include step="0.01" and
modified the JavaScript parsing logic to use parseFloat instead of
parseInt. This allows users to enter decimal values like 10.6 for
product quantities without triggering validation errors or truncation.

+++ orders.js
@@ -79,7 +79,7 @@ const handleFormSubmit = async (e) => {
         date: orderForm['order-date'].value,
         customerId: orderForm['order-customer'].value,
         productId: orderForm['order-product'].value,
-        quantity: parseInt(orderForm['order-quantity'].value),
+        quantity: parseFloat(orderForm['order-quantity'].value),
         total: parseFloat(orderForm['order-total'].value),
         paid: parseFloat(orderForm['order-paid'].value),
         status: orderForm['order-status'].value,
